plugin/sign: a plugin that signs zone
Sign is a plugin that signs zone data (on disk). The README.md details
what exactly happens to should be accurate related to the code.

There are a couple of things missing, documented in TODO; these mostly
deal with setup and tear down. But also some more important things like
adding NSEC and not signing glue records. Also more tests need be added.

The current test signs db.miek.nl and puts the result in
db.miek.nl.signed in the current directory.

The change to file/* need to be backed out, once some other changes in
another PR are in - these are just here to make things work now.

Signed-off-by: Miek Gieben <miek@miek.nl>